This commit contains changes in fseg_filters.py, which now contains overlay function and the replace of all math functions to numpy functions. Also, for now it's saving the final results in a paste "ResultImages" -> See .gitignore for more details
