+++ .gitignore
@@ -1,2 +1,3 @@
 /Images/
 /.idea/
+/ResultImages/
