+++ FctSeg_seed.py
@@ -5,12 +5,11 @@ Factorization based segmentation with manually selected seeds
 """
 import time
 import numpy as np
-from fseg_filters import image_filtering
-import matplotlib.pyplot as plt
+import argparse
+import cv2
 
 from scipy import linalg as LAsci
-from skimage import io
-
+from utils.fseg_filters import image_filtering, overlay
 
 def SHcomp(Ig, ws, BinN=11):
     """
@@ -105,9 +104,23 @@ def Fseg(Ig, ws, seeds):
 
 
 if __name__ == '__main__':
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument("-f", "--file", help="file path")
+    parser.add_argument("-ws", "--window_size", help="window size for local special histogram")
+    # parser.add_argument("-s", "--seeds",
+    #                     help="list of coordinates [row, column] for seeds. each seed represent one type of texture")
+    args = parser.parse_args()
+
+    file_path = args.file
+    ws = int(args.window_size)
+    #seeds = args.seeds
+    # n_segments = int(args.number_segments)
+    # omega = float(args.error_treshold)
+    # nonneg_constraint = bool(args.nonneg_constraint)
     time0 = time.time()
     # an example of using Fseg
-    img = io.imread('../M1.pgm')
+    img = cv2.imread(file_path, 0)
 
     # define filter bank and apply to image. for color images, convert rgb to grey scale and then apply filter bank
     filter_list = [('log', .5, [3, 3]), ('log', 1.2, [7, 7])]
@@ -120,15 +133,16 @@ if __name__ == '__main__':
     seeds = [[60, 238], [160, 160], [238, 60]]  # provide seeds
 
     # run segmentation. try different window size
-    seg_out = Fseg(Ig, ws=19, seeds=seeds)
+    seg_out = Fseg(Ig, ws=ws, seeds=seeds)
 
     print('FSEG runs in %0.2f seconds. ' % (time.time() - time0))
 
     # show results
-    fig, ax = plt.subplots(ncols=2, sharex=True, sharey=True, figsize=(10, 5))
-    ax[0].imshow(img, cmap='gray')
-    seeds = np.array(seeds)
-    plt.plot(seeds[:, 1], seeds[:, 0], 'r*')
-    ax[1].imshow(seg_out, cmap='gray')
-    plt.tight_layout()
-    plt.show()
+    # fig, ax = plt.subplots(ncols=2, sharex=True, sharey=True, figsize=(10, 5))
+    # ax[0].imshow(img, cmap='gray')
+    # seeds = np.array(seeds)
+    # plt.plot(seeds[:, 1], seeds[:, 0], 'r*')
+    # ax[1].imshow(seg_out, cmap='gray')
+    # plt.tight_layout()
+    # plt.show()
+    overlay(img, seg_out, 0.6, cmap="viridis")
